Final updates/backend/requirements.txt
django==4.2.20
firebase-admin==6.6.0
python-dotenv==1.0.1
django-cors-headers==4.7.0
requests==2.31.0 